kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/fast/events/touch/touch-transform-element.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="./resources/touch-util.js"></script>
<style>
.test {
    width: 100px;
    height: 100px;
    background-color: green;
}

#test0 {
    position: absolute;
    top: 0;
    left: 50px;
    width:111;
}
#test1 {
    position: absolute;
    top: 100px;
    left: 100px;
    width: 202px;
    height: 200px;
    -webkit-transform: translate3d(100px, 100px, 0px) scale(0.5);
}

</style>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<div class="test" id="test0"></div>
<div class="test" id="test1"></div>

<script type="text/javascript">
var lastEvent = null;
var touchEventsReceived = 0;
var EXPECTED_TOUCH_EVENTS_TOTAL = 12;

function touchEventCallback(event) {
    if (window.eventSender) {
        lastEvent = event;
        verifyTouch(touchEventsReceived++);
    } else {
        debug(event.type);
    }

    if (window.layoutTestController && touchEventsReceived == EXPECTED_TOUCH_EVENTS_TOTAL) {
        // If we've got here, we can safely say we were successfully parsed :) We need to
        // call the isSucccessfullyParsed function to output the correct TEST COMPLETE
        // footer message.
        isSuccessfullyParsed();
        layoutTestController.notifyDone();
    }
}

var divs = document.getElementsByClassName("test");
for (var i = 0; i < divs.length; ++i) {
  var div = divs[i];
  div.addEventListener("touchstart", touchEventCallback, false);
  div.addEventListener("touchmove", touchEventCallback, false);
  div.addEventListener("touchend", touchEventCallback, false);
}

function verifyTouchEvent(type, totalTouchCount, changedTouchCount, targetTouchCount)
{
    shouldBeEqualToString("lastEvent.type", type);
    shouldBe("lastEvent.touches.length", totalTouchCount.toString());
    shouldBe("lastEvent.changedTouches.length", changedTouchCount.toString());
    shouldBe("lastEvent.targetTouches.length", targetTouchCount.toString());
    shouldBe("lastEvent.pageX", "0");
    shouldBe("lastEvent.pageY", "0");
}

function verifyTouchPoint(list, point, x, y, id)
{
    shouldBe("lastEvent." + list + "[" + point + "].pageX", x.toString());
    shouldBe("lastEvent." + list + "[" + point + "].pageY", y.toString());
    shouldBe("lastEvent." + list + "[" + point + "].clientX", x.toString());
    shouldBe("lastEvent." + list + "[" + point + "].clientY", y.toString());
    shouldBe("lastEvent." + list + "[" + point + "].identifier", id.toString());
}

function verifyTouch(which) {

    switch (which) {
    case 0:
        verifyTouchEvent("touchstart", 1, 1, 1);
        shouldBeEqualToString("lastEvent.touches[0].target.id", "test0");
        verifyTouchPoint("touches", 0, 100, 50, 0);
        verifyTouchPoint("changedTouches", 0, 100, 50, 0);
        verifyTouchPoint("targetTouches", 0, 100, 50, 0);
        break;
    case 1:
        verifyTouchEvent("touchmove", 1, 1, 1);
        verifyTouchPoint("touches", 0, 50, 50, 0);
        break;
    case 2:
        verifyTouchEvent("touchend", 0, 1, 0);
        verifyTouchPoint("changedTouches", 0, 50, 50, 0);
        break;

    case 3:
        verifyTouchEvent("touchstart", 1, 1, 1);
        shouldBeEqualToString("lastEvent.targetTouches[0].target.id", "test1");
        verifyTouchPoint("touches", 0, 300, 300, 0);
        verifyTouchPoint("changedTouches", 0, 300, 300, 0);
        verifyTouchPoint("targetTouches", 0, 300, 300, 0);
        break;
    case 4:
        verifyTouchEvent("touchmove", 1, 1, 1);
        verifyTouchPoint("touches", 0, 1000, 1000, 0);
        break;
    case 5:
        verifyTouchEvent("touchend", 0, 1, 0);
        verifyTouchPoint("changedTouches", 0, 1000, 1000, 0);
        break;

    case 6:
        verifyTouchEvent("touchstart", 1, 1, 1);
        shouldBeEqualToString("lastEvent.touches[0].target.id", "test0");
        verifyTouchPoint("touches", 0, 25, 175, 0);
        verifyTouchPoint("changedTouches", 0, 25, 175, 0);
        verifyTouchPoint("targetTouches", 0, 25, 175, 0);
        break;
    case 7:
        verifyTouchEvent("touchmove", 1, 1, 1);
        verifyTouchPoint("touches", 0, 75, 75, 0);
        break;
    case 8:
        verifyTouchEvent("touchend", 0, 1, 0);
        verifyTouchPoint("changedTouches", 0, 75, 75, 0);
        break;

    case 9:
        verifyTouchEvent("touchstart", 1, 1, 1);
        shouldBeEqualToString("lastEvent.targetTouches[0].target.id", "test1");
        verifyTouchPoint("touches", 0, 476, 476, 0);
        verifyTouchPoint("changedTouches", 0, 476, 476, 0);
        verifyTouchPoint("targetTouches", 0, 476, 476, 0);
        break;
    case 10:
        verifyTouchEvent("touchmove", 1, 1, 1);
        verifyTouchPoint("touches", 0, 300, 300, 0);
        break;
    case 11:
        verifyTouchEvent("touchend", 0, 1, 0);
        verifyTouchPoint("changedTouches", 0, 300, 300, 0);
        break;

        default: testFailed("Wrong number of touch events! (" + which + ")");
    }
}

function touchStaticElement1() {
    eventSender.addTouchPoint(100, 50);
    eventSender.touchStart();

    eventSender.updateTouchPoint(0, 50, 50);
    eventSender.touchMove();

    eventSender.releaseTouchPoint(0);
    eventSender.touchEnd();
}

function touchTransformedElement1() {
    eventSender.addTouchPoint(300, 300);
    eventSender.touchStart();

    eventSender.updateTouchPoint(0, 1000, 1000);
    eventSender.touchMove();

    eventSender.releaseTouchPoint(0);
    eventSender.touchEnd();
}

function moveStaticElement() {
    document.getElementById("test0").style.top = "150px";
    document.getElementById("test0").style.left = "0px";
}

function moveTransformedElement() {
   document.getElementById("test1").style.webkitTransform = "translate3d(300px, 300px, 0px) scale(0.25)"
}


function touchStaticElement2() {
    eventSender.addTouchPoint(25, 175);
    eventSender.touchStart();

    eventSender.updateTouchPoint(0, 75, 75);
    eventSender.touchMove();

    eventSender.releaseTouchPoint(0);
    eventSender.touchEnd();
}

function touchTransformedElement2() {
    eventSender.addTouchPoint(476, 476);
    eventSender.touchStart();

    eventSender.updateTouchPoint(0, 300, 300);
    eventSender.touchMove();

    eventSender.releaseTouchPoint(0);
    eventSender.touchEnd();
}

if (window.layoutTestController)
    layoutTestController.waitUntilDone();

runTouchTest(function() {
    lastEvent = null;
    var scriptId = 0;
    var script = [
        touchStaticElement1,
        touchTransformedElement1,
        moveStaticElement,
        touchStaticElement1, // Shouldn't cause anything.
        touchStaticElement2,
        moveTransformedElement,
        touchTransformedElement1, // Shouldn't cause anything.
        touchTransformedElement2
    ];
    function runScript() {
        if (typeof (script[scriptId]) === "function") {
            (script[scriptId++])();
            setTimeout(runScript, 50);
        }
    }
    runScript();
});

</script>
</body>
</html>
