kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/platform/android/fast/events/touch/touch-transform-element-expected.txt

PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].target.id is "test0"
PASS lastEvent.touches[0].pageX is 100
PASS lastEvent.touches[0].pageY is 50
PASS lastEvent.touches[0].clientX is 100
PASS lastEvent.touches[0].clientY is 50
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.changedTouches[0].pageX is 100
PASS lastEvent.changedTouches[0].pageY is 50
PASS lastEvent.changedTouches[0].clientX is 100
PASS lastEvent.changedTouches[0].clientY is 50
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.targetTouches[0].pageX is 100
PASS lastEvent.targetTouches[0].pageY is 50
PASS lastEvent.targetTouches[0].clientX is 100
PASS lastEvent.targetTouches[0].clientY is 50
PASS lastEvent.targetTouches[0].identifier is 0
PASS lastEvent.type is "touchmove"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].pageX is 50
PASS lastEvent.touches[0].pageY is 50
PASS lastEvent.touches[0].clientX is 50
PASS lastEvent.touches[0].clientY is 50
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.type is "touchend"
PASS lastEvent.touches.length is 0
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 0
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.changedTouches[0].pageX is 50
PASS lastEvent.changedTouches[0].pageY is 50
PASS lastEvent.changedTouches[0].clientX is 50
PASS lastEvent.changedTouches[0].clientY is 50
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.targetTouches[0].target.id is "test1"
PASS lastEvent.touches[0].pageX is 300
PASS lastEvent.touches[0].pageY is 300
PASS lastEvent.touches[0].clientX is 300
PASS lastEvent.touches[0].clientY is 300
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.changedTouches[0].pageX is 300
PASS lastEvent.changedTouches[0].pageY is 300
PASS lastEvent.changedTouches[0].clientX is 300
PASS lastEvent.changedTouches[0].clientY is 300
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.targetTouches[0].pageX is 300
PASS lastEvent.targetTouches[0].pageY is 300
PASS lastEvent.targetTouches[0].clientX is 300
PASS lastEvent.targetTouches[0].clientY is 300
PASS lastEvent.targetTouches[0].identifier is 0
PASS lastEvent.type is "touchmove"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].pageX is 1000
PASS lastEvent.touches[0].pageY is 1000
PASS lastEvent.touches[0].clientX is 1000
PASS lastEvent.touches[0].clientY is 1000
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.type is "touchend"
PASS lastEvent.touches.length is 0
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 0
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.changedTouches[0].pageX is 1000
PASS lastEvent.changedTouches[0].pageY is 1000
PASS lastEvent.changedTouches[0].clientX is 1000
PASS lastEvent.changedTouches[0].clientY is 1000
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].target.id is "test0"
PASS lastEvent.touches[0].pageX is 25
PASS lastEvent.touches[0].pageY is 175
PASS lastEvent.touches[0].clientX is 25
PASS lastEvent.touches[0].clientY is 175
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.changedTouches[0].pageX is 25
PASS lastEvent.changedTouches[0].pageY is 175
PASS lastEvent.changedTouches[0].clientX is 25
PASS lastEvent.changedTouches[0].clientY is 175
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.targetTouches[0].pageX is 25
PASS lastEvent.targetTouches[0].pageY is 175
PASS lastEvent.targetTouches[0].clientX is 25
PASS lastEvent.targetTouches[0].clientY is 175
PASS lastEvent.targetTouches[0].identifier is 0
PASS lastEvent.type is "touchmove"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].pageX is 75
PASS lastEvent.touches[0].pageY is 75
PASS lastEvent.touches[0].clientX is 75
PASS lastEvent.touches[0].clientY is 75
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.type is "touchend"
PASS lastEvent.touches.length is 0
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 0
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.changedTouches[0].pageX is 75
PASS lastEvent.changedTouches[0].pageY is 75
PASS lastEvent.changedTouches[0].clientX is 75
PASS lastEvent.changedTouches[0].clientY is 75
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.targetTouches[0].target.id is "test1"
PASS lastEvent.touches[0].pageX is 476
PASS lastEvent.touches[0].pageY is 476
PASS lastEvent.touches[0].clientX is 476
PASS lastEvent.touches[0].clientY is 476
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.changedTouches[0].pageX is 476
PASS lastEvent.changedTouches[0].pageY is 476
PASS lastEvent.changedTouches[0].clientX is 476
PASS lastEvent.changedTouches[0].clientY is 476
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.targetTouches[0].pageX is 476
PASS lastEvent.targetTouches[0].pageY is 476
PASS lastEvent.targetTouches[0].clientX is 476
PASS lastEvent.targetTouches[0].clientY is 476
PASS lastEvent.targetTouches[0].identifier is 0
PASS lastEvent.type is "touchmove"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.touches[0].pageX is 300
PASS lastEvent.touches[0].pageY is 300
PASS lastEvent.touches[0].clientX is 300
PASS lastEvent.touches[0].clientY is 300
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.type is "touchend"
PASS lastEvent.touches.length is 0
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 0
PASS lastEvent.pageX is 0
PASS lastEvent.pageY is 0
PASS lastEvent.changedTouches[0].pageX is 300
PASS lastEvent.changedTouches[0].pageY is 300
PASS lastEvent.changedTouches[0].clientX is 300
PASS lastEvent.changedTouches[0].clientY is 300
PASS lastEvent.changedTouches[0].identifier is 0
PASS successfullyParsed is true

TEST COMPLETE
